free operations from rapids
