+++ h2o-core/src/main/java/water/rapids/ast/prims/math/AstGamma.java
@@ -1,6 +1,7 @@
 package water.rapids.ast.prims.math;
 
-import org.apache.commons.math3.special.Gamma;
+
+import water.operations.Unary;
 
 /**
  */
@@ -12,6 +13,6 @@ public class AstGamma extends AstUniOp {
 
   @Override
   public double op(double d) {
-    return Gamma.gamma(d);
+      return Unary.gamma(d);
   }
 }
