+++ h2o-core/src/main/java/water/rapids/ast/prims/math/AstSignif.java
@@ -1,5 +1,6 @@
 package water.rapids.ast.prims.math;
 
+import water.operations.Unary;
 import water.rapids.ast.prims.operators.AstBinOp;
 
 /**
@@ -10,11 +11,7 @@ public class AstSignif extends AstBinOp {
   }
 
   public double op(double x, double digits) {
-    if (Double.isNaN(x)) return x;
-    if (digits < 1) digits = 1; //mimic R's base::signif
-    if ((int) digits != digits) digits = Math.round(digits);
-    java.math.BigDecimal bd = new java.math.BigDecimal(x);
-    bd = bd.round(new java.math.MathContext((int) digits, java.math.RoundingMode.HALF_EVEN));
-    return bd.doubleValue();
+      return Unary.signif(x, digits);
   }
+
 }
