+++ h2o-core/src/main/java/water/rapids/ast/prims/math/AstSinPi.java
@@ -2,6 +2,7 @@ package water.rapids.ast.prims.math;
 
 /**
  */
+@Deprecated
 public class AstSinPi extends AstUniOp {
   @Override
   public String str() {
