+++ h2o-core/src/main/java/water/rapids/ast/prims/math/AstTriGamma.java
@@ -1,6 +1,6 @@
 package water.rapids.ast.prims.math;
 
-import org.apache.commons.math3.special.Gamma;
+import water.operations.Unary;
 
 /**
  */
@@ -12,6 +12,7 @@ public class AstTriGamma extends AstUniOp {
 
   @Override
   public double op(double d) {
-    return Double.isNaN(d) ? Double.NaN : Gamma.trigamma(d);
+      return Unary.trigamma(d);
   }
+
 }
