+++ h2o-core/src/main/java/water/rapids/ast/prims/mungers/AstAsCharacter.java
@@ -2,8 +2,8 @@ package water.rapids.ast.prims.mungers;
 
 import water.fvec.Frame;
 import water.fvec.Vec;
+import water.operations.FrameUtils;
 import water.rapids.Env;
-import water.rapids.Val;
 import water.rapids.ast.AstRoot;
 import water.rapids.vals.ValFrame;
 import water.rapids.ast.AstPrimitive;
@@ -31,17 +31,8 @@ public class AstAsCharacter extends AstPrimitive {
   @Override
   public ValFrame apply(Env env, Env.StackHelp stk, AstRoot asts[]) {
     Frame ary = stk.track(asts[1].exec(env)).getFrame();
-    Vec[] nvecs = new Vec[ary.numCols()];
-    Vec vv;
-    for (int c = 0; c < nvecs.length; ++c) {
-      vv = ary.vec(c);
-      try {
-        nvecs[c] = vv.toStringVec();
-      } catch (Exception e) {
-        VecUtils.deleteVecs(nvecs, c);
-        throw e;
-      }
-    }
-    return new ValFrame(new Frame(ary._names, nvecs));
+      Frame frame = FrameUtils.convertAllColumnsToString(ary);
+      return new ValFrame(frame);
   }
+
 }
