+++ h2o-core/src/main/java/water/rapids/ast/prims/mungers/AstLevels.java
@@ -31,32 +31,36 @@ public class AstLevels extends AstPrimitive {
 
   @Override
   public ValFrame apply(Env env, Env.StackHelp stk, AstRoot asts[]) {
-    Frame f = stk.track(asts[1].exec(env)).getFrame();
-    Futures fs = new Futures();
-    Key[] keys = Vec.VectorGroup.VG_LEN1.addVecs(f.numCols());
-    Vec[] vecs = new Vec[keys.length];
-
-    // compute the longest vec... that's the one with the most domain levels
-    int max = 0;
-    for (int i = 0; i < f.numCols(); ++i)
-      if (f.vec(i).isCategorical())
-        if (max < f.vec(i).domain().length) max = f.vec(i).domain().length;
-
-    final int rowLayout = Vec.ESPC.rowLayout(keys[0], new long[]{0, max});
-    for (int i = 0; i < f.numCols(); ++i) {
-      AppendableVec v = new AppendableVec(keys[i], Vec.T_NUM);
-      NewChunk nc = new NewChunk(v, 0);
-      String[] dom = f.vec(i).domain();
-      int numToPad = dom == null ? max : max - dom.length;
-      if (dom != null)
-        for (int j = 0; j < dom.length; ++j) nc.addNum(j);
-      for (int j = 0; j < numToPad; ++j) nc.addNA();
-      nc.close(0, fs);
-      vecs[i] = v.close(rowLayout, fs);
-      vecs[i].setDomain(dom);
-    }
-    fs.blockForPending();
-    Frame fr2 = new Frame(vecs);
-    return new ValFrame(fr2);
+      Frame f = stk.track(asts[1].exec(env)).getFrame();
+      Frame fr2 = levels(f);
+      return new ValFrame(fr2);
   }
+
+    public Frame levels(Frame f) {
+        Futures fs = new Futures();
+        Key[] keys = Vec.VectorGroup.VG_LEN1.addVecs(f.numCols());
+        Vec[] vecs = new Vec[keys.length];
+
+        // compute the longest vec... that's the one with the most domain levels
+        int max = 0;
+        for (int i = 0; i < f.numCols(); ++i)
+            if (f.vec(i).isCategorical())
+                if (max < f.vec(i).domain().length) max = f.vec(i).domain().length;
+
+        final int rowLayout = Vec.ESPC.rowLayout(keys[0], new long[]{0, max});
+        for (int i = 0; i < f.numCols(); ++i) {
+            AppendableVec v = new AppendableVec(keys[i], Vec.T_NUM);
+            NewChunk nc = new NewChunk(v, 0);
+            String[] dom = f.vec(i).domain();
+            int numToPad = dom == null ? max : max - dom.length;
+            if (dom != null)
+                for (int j = 0; j < dom.length; ++j) nc.addNum(j);
+            for (int j = 0; j < numToPad; ++j) nc.addNA();
+            nc.close(0, fs);
+            vecs[i] = v.close(rowLayout, fs);
+            vecs[i].setDomain(dom);
+        }
+        fs.blockForPending();
+        return new Frame(vecs);
+    }
 }
