+++ h2o-core/src/main/java/water/rapids/ast/prims/reducers/AstMean.java
@@ -1,8 +1,8 @@
 package water.rapids.ast.prims.reducers;
 
-import water.Key;
-import water.MRTask;
 import water.fvec.*;
+import water.operations.ColumnWise;
+import water.operations.RowWise;
 import water.rapids.Env;
 import water.rapids.Val;
 import water.rapids.vals.ValFrame;
@@ -55,7 +55,7 @@ public class AstMean extends AstPrimitive {
       Frame fr = stk.track(val1).getFrame();
       boolean na_rm = asts[2].exec(env).getNum() == 1;
       boolean axis = asts.length == 4 && (asts[3].exec(env).getNum() == 1);
-      return axis? rowwiseMean(fr, na_rm) : colwiseMean(fr, na_rm);
+        return axis ? RowWise.mean(fr, na_rm) : ColumnWise.mean(fr, na_rm);
     }
     else if (val1 instanceof ValRow) {
       // This may be called from AstApply when doing per-row computations.
@@ -78,94 +78,4 @@ public class AstMean extends AstPrimitive {
   }
 
 
-  /**
-   * Compute Frame means by rows, and return a frame consisting of a single Vec of means in each row.
-   */
-  private ValFrame rowwiseMean(Frame fr, final boolean na_rm) {
-    String[] newnames = {"mean"};
-    Key<Frame> newkey = Key.make();
-
-    // Determine how many columns of different types we have
-    int n_numeric = 0, n_time = 0;
-    for (Vec vec : fr.vecs()) {
-      if (vec.isNumeric()) n_numeric++;
-      if (vec.isTime()) n_time++;
-    }
-    // Compute the type of the resulting column: if all columns are TIME then the result is also time; otherwise
-    // if at least one column is numeric then the result is also numeric.
-    byte resType = n_numeric > 0? Vec.T_NUM : Vec.T_TIME;
-
-    // Construct the frame over which the mean should be computed
-    Frame compFrame = new Frame();
-    for (int i = 0; i < fr.numCols(); i++) {
-      Vec vec = fr.vec(i);
-      if (n_numeric > 0? vec.isNumeric() : vec.isTime())
-        compFrame.add(fr.name(i), vec);
-    }
-    Vec anyvec = compFrame.anyVec();
-
-    // Take into account certain corner cases
-    if (anyvec == null) {
-      Frame res = new Frame(newkey);
-      anyvec = fr.anyVec();
-      if (anyvec != null) {
-        // All columns in the original frame are non-numeric -> return a vec of NAs
-        res.add("mean", anyvec.makeCon(Double.NaN));
-      } // else the original frame is empty, in which case we return an empty frame too
-      return new ValFrame(res);
-    }
-    if (!na_rm && n_numeric < fr.numCols() && n_time < fr.numCols()) {
-      // If some of the columns are non-numeric and na_rm==false, then the result is a vec of NAs
-      Frame res = new Frame(newkey, newnames, new Vec[]{anyvec.makeCon(Double.NaN)});
-      return new ValFrame(res);
-    }
-
-    // Compute the mean over all rows
-    final int numCols = compFrame.numCols();
-    Frame res = new MRTask() {
-      @Override
-      public void map(Chunk[] cs, NewChunk nc) {
-        for (int i = 0; i < cs[0]._len; i++) {
-          double d = 0;
-          int numNaColumns = 0;
-          for (int j = 0; j < numCols; j++) {
-            double val = cs[j].atd(i);
-            if (Double.isNaN(val))
-              numNaColumns++;
-            else
-              d += val;
-          }
-          if (na_rm? numNaColumns < numCols : numNaColumns == 0)
-            nc.addNum(d / (numCols - numNaColumns));
-          else
-            nc.addNum(Double.NaN);
-        }
-      }
-    }.doAll(1, resType, compFrame)
-     .outputFrame(newkey, newnames, null);
-
-    // Return the result
-    return new ValFrame(res);
-  }
-
-
-  /**
-   * Compute column-wise means (i.e. means of each column), and return a frame having a single row.
-   */
-  private ValFrame colwiseMean(Frame fr, final boolean na_rm) {
-    Frame res = new Frame();
-
-    Vec vec1 = Vec.makeCon(null, 0);
-    assert vec1.length() == 1;
-
-    for (int i = 0; i < fr.numCols(); i++) {
-      Vec v = fr.vec(i);
-      boolean valid = (v.isNumeric() || v.isTime() || v.isBinary()) && v.length() > 0 && (na_rm || v.naCnt() == 0);
-      Vec newvec = vec1.makeCon(valid? v.mean() : Double.NaN, v.isTime()? Vec.T_TIME : Vec.T_NUM);
-      res.add(fr.name(i), newvec);
-    }
-
-    vec1.remove();
-    return new ValFrame(res);
-  }
 }
